requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
atomicwrites==1.4.1
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest_check==1.0.5
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.11
